Version 0.3.0.3 Se arreglaron las bajas de prestamo y los listados de prestamo

+++ Persistencia/pPrestamoEquipo.cs
@@ -117,12 +117,80 @@ namespace Persistencia
 
         public int bajaPrestamoDeEquipo(int idPrestamo)
         {
+            int id1 = 0;
+            int id2 = 0;
+            int id3 = 0;
+            int id5 = 0;
+            string consultaSQL2;
+            string consultaSQL = "SELECT * FROM prestamoDeEquipo WHERE prestamoDeEquipo.id_Prestamo = " + idPrestamo + ";";
+            MySqlDataReader fila = ejecutarYdevolver(consultaSQL);
+            while (fila.Read())
+            {
 
-            string consultaSQL = "DELETE FROM prestamoDeEquipo WHERE prestamoDeEquipo.id_Prestamo = " + idPrestamo + " ;";
-            ejecutarSQL(consultaSQL);
+                id1 = (idPrestamo);
+                id1 = recrearIdPrestamo(fila);
+
+            }
+            if (id1 != 0)
+            {
+
+                consultaSQL2 = "DELETE FROM prestamoDeEquipo WHERE prestamoDeEquipo.id_Prestamo = " + idPrestamo + " ;";
+                ejecutarSQL(consultaSQL);
+
+            }
+
+            consultaSQL = "SELECT * FROM prestamoDeEspacio WHERE prestamoDeEspacio.id_Prestamo = " + idPrestamo + ";";
+            MySqlDataReader fila2 = ejecutarYdevolver(consultaSQL);
+            while (fila2.Read())
+            {
+
+                id2 = (idPrestamo);
+                id2 = recrearIdPrestamo(fila2);
+
+            }
+            if (id2 != 0)
+            {
+
+                consultaSQL2 = "DELETE FROM prestamoDeEspacio WHERE prestamoDeEspacio.id_Prestamo = " + idPrestamo + ";";
+                ejecutarSQL(consultaSQL2);
+
+            }
+
+            consultaSQL = "SELECT * FROM prestamoEspontaneo WHERE prestamoEspontaneo.id_Prestamo = " + idPrestamo + ";";
+            MySqlDataReader fila3 = ejecutarYdevolver(consultaSQL);
+            while (fila3.Read())
+            {
+
+                id3 = (idPrestamo);
+                id3 = recrearIdPrestamo(fila3);
+
+            }
+            if (id3 != 0)
+            {
+
+                consultaSQL2 = "DELETE FROM prestamoEspontaneo WHERE prestamoEspontaneo.id_Prestamo = " + idPrestamo + ";";
+                ejecutarSQL(consultaSQL2);
+                
+            }
+
+            consultaSQL = "SELECT * FROM prestamo WHERE prestamo.id = " + idPrestamo + ";";
+            MySqlDataReader fila5 = ejecutarYdevolver(consultaSQL);
+            while (fila5.Read())
+            {
+
+                id5 = (idPrestamo);
+                id5 = recrearIdPrestamo(fila5);
+
+            }
+            if (id5 != 0)
+            {
+
+                consultaSQL2 = "DELETE FROM prestamo WHERE prestamo.id_Prestamo = " + idPrestamo + ";";
+                ejecutarSQL(consultaSQL2);
+
+            }
             
             return idPrestamo;
-
         }
 
         public bool modificacionPrestamo(ePrestamoEquipo unEPE, String IdPrestamo)
